tmp commit

+++ src/transformers/models/llama/modeling_llama.py
@@ -150,8 +150,8 @@ class LlamaRotaryEmbedding(nn.Module):
             self._set_cos_sin_cache(seq_len=seq_len, device=x.device, dtype=x.dtype)
 
         return (
-            self.cos_cached[:seq_len].to(x.device),
-            self.sin_cached[:seq_len].to(x.device),
+            self.cos_cached[:seq_len].to(x.device).to(x.dtype),
+            self.sin_cached[:seq_len].to(x.device).to(x.dtype),
         )
